requirements.txt
paddleocr==2.6.1.3
paddlepaddle-gpu==2.5.2
flask
flask-cors
pymongo
pillow
requests
python-dotenv

